requirements.txt
streamlit
requests
PyPDF2
Pillow
SpeechRecognition
pyaudio